# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/navigational-list/auth-test": {
      "get": {
        "operationId": "MenuController_authTest",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTestDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list": {
      "post": {
        "operationId": "MenuController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      },
      "get": {
        "operationId": "MenuController_findAll",
        "parameters": [
          {
            "name": "domain",
            "required": false,
            "in": "query",
            "description": "Filter by domain",
            "schema": {
              "enum": [
                "ADMIN",
                "WORKSHOP"
              ],
              "type": "string"
            }
          },
          {
            "name": "structuralSubtype",
            "required": false,
            "in": "query",
            "description": "Filter by structural subtype",
            "schema": {
              "enum": [
                "HEADER",
                "NAV",
                "FOOTER"
              ],
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "description": "Filter by state",
            "schema": {
              "enum": [
                "FULL",
                "RELAXED",
                "COMPACT"
              ],
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "description": "Filter by archived status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "authRequired",
            "required": false,
            "in": "query",
            "description": "Filter by authentication requirement",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/hierarchy/{domain}": {
      "get": {
        "operationId": "MenuController_getMenuHierarchy",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "ADMIN",
                "WORKSHOP"
              ],
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived menu items",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete menu hierarchy for the domain"
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/domain/{domain}": {
      "get": {
        "operationId": "MenuController_findByDomain",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "ADMIN",
                "WORKSHOP"
              ],
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived menu items",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/domain/{domain}/structural-subtype/{structuralSubtype}": {
      "get": {
        "operationId": "MenuController_findByDomainAndStructuralSubtype",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "ADMIN",
                "WORKSHOP"
              ],
              "type": "string"
            }
          },
          {
            "name": "structuralSubtype",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "HEADER",
                "NAV",
                "FOOTER"
              ],
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived menu items",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/domain/{domain}/structural-subtype/{structuralSubtype}/state/{state}": {
      "get": {
        "operationId": "MenuController_findByDomainStructuralSubtypeAndState",
        "parameters": [
          {
            "name": "domain",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "ADMIN",
                "WORKSHOP"
              ],
              "type": "string"
            }
          },
          {
            "name": "structuralSubtype",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "HEADER",
                "NAV",
                "FOOTER"
              ],
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "FULL",
                "RELAXED",
                "COMPACT"
              ],
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived menu items",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItemDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/sort/{id}": {
      "post": {
        "operationId": "MenuController_reorderMenuItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Menu item ID to reposition",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MenuItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/{id}": {
      "get": {
        "operationId": "MenuController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      },
      "patch": {
        "operationId": "MenuController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      },
      "delete": {
        "operationId": "MenuController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/{id}/archive": {
      "patch": {
        "operationId": "MenuController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    },
    "/navigational-list/{id}/unarchive": {
      "patch": {
        "operationId": "MenuController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemDto"
                }
              }
            }
          }
        },
        "tags": [
          "Menu"
        ]
      }
    }
  },
  "info": {
    "title": "NGX service-navigational-list API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "AuthTestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateMenuItemDto": {
        "type": "object",
        "properties": {
          "menuItemText": {
            "type": "string"
          },
          "routePath": {
            "type": "string"
          },
          "tooltipText": {
            "type": "string"
          },
          "navSvgPath": {
            "type": "string"
          },
          "headerSvgPath": {
            "type": "string"
          },
          "sortId": {
            "type": "number"
          },
          "authRequired": {
            "type": "boolean"
          },
          "domain": {
            "type": "string",
            "enum": [
              "ADMIN",
              "WORKSHOP"
            ]
          },
          "structuralSubtype": {
            "type": "string",
            "enum": [
              "HEADER",
              "NAV",
              "FOOTER"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "FULL",
              "RELAXED",
              "COMPACT"
            ]
          },
          "description": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string",
            "description": "Parent menu item ID for hierarchical navigation"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "menuItemText",
          "routePath",
          "sortId",
          "domain",
          "structuralSubtype",
          "state"
        ]
      },
      "MenuItemDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "menuItemText": {
            "type": "string"
          },
          "routePath": {
            "type": "string"
          },
          "tooltipText": {
            "type": "string"
          },
          "navSvgPath": {
            "type": "string"
          },
          "headerSvgPath": {
            "type": "string"
          },
          "sortId": {
            "type": "number"
          },
          "authRequired": {
            "type": "boolean"
          },
          "domain": {
            "type": "string",
            "enum": [
              "ADMIN",
              "WORKSHOP"
            ]
          },
          "structuralSubtype": {
            "type": "string",
            "enum": [
              "HEADER",
              "NAV",
              "FOOTER"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "FULL",
              "RELAXED",
              "COMPACT"
            ]
          },
          "version": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          },
          "__v": {
            "type": "number"
          },
          "parentId": {
            "type": "string",
            "description": "Parent menu item ID for hierarchical navigation"
          }
        },
        "required": [
          "_id",
          "menuItemText",
          "routePath",
          "sortId",
          "authRequired",
          "domain",
          "structuralSubtype",
          "state",
          "version",
          "description",
          "lastUpdated",
          "archived",
          "__v"
        ]
      },
      "UpdateMenuItemDto": {
        "type": "object",
        "properties": {
          "menuItemText": {
            "type": "string"
          },
          "routePath": {
            "type": "string"
          },
          "tooltipText": {
            "type": "string"
          },
          "navSvgPath": {
            "type": "string"
          },
          "headerSvgPath": {
            "type": "string"
          },
          "sortId": {
            "type": "number"
          },
          "authRequired": {
            "type": "boolean"
          },
          "domain": {
            "type": "string",
            "enum": [
              "ADMIN",
              "WORKSHOP"
            ]
          },
          "structuralSubtype": {
            "type": "string",
            "enum": [
              "HEADER",
              "NAV",
              "FOOTER"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "FULL",
              "RELAXED",
              "COMPACT"
            ]
          },
          "description": {
            "type": "string"
          },
          "archived": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string",
            "description": "Parent menu item ID for hierarchical navigation"
          },
          "lastUpdated": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}